requirements.txt
streamlit==1.26.0
tensorflow==2.12.1
Pillow
numpy
requests
